feat: support binding refs directly to components
It was previously assumed that the consumer always need to pass a a raw Element or Ref<Element>. This means that when assigning this refs to Components they would have to use the function ref notation and extract the element from ComponentPublicInstance.
This increases the package size but makes it more convenient. I'm happy to provide util instead calling it internally to make the package smaller, but this makes the api a bit more intuitive so i'll ship it for now.

Before
<some-comp :ref=(cpi) => reference = cpi.$el" />
After
<some-comp ref=reference" />

https://vuejs.org/guide/essentials/template-refs.html#ref-on-component

Update utils.js